requirements.txt
torch>=1.9.0
torch-geometric>=2.0.0
pytorch-lightning>=1.8.0
numpy>=1.19.0
pandas>=1.2.0
scipy>=1.6.0
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.60.0
scikit-learn>=0.24.0
tensorboard>=2.7.0
hydra-core>=1.1.0
omegaconf>=2.1.0